+++ Changelog.md
@@ -10,6 +10,8 @@
 - Link the `Concentration` status to the effects transferred to Actors, so when the `Concentration` is cancelled then the transferred effects are removed from the targeted tokens.
 - "Year 1 Pan Pacifica" game system option wording changed, and when checked the spirit axiom for Pan Pacifica is changed to the new value.
 - Shorter "end encounter" button for our French users.
+- Allow any skill to be specified in an Active Effect's `Apply If Attack/Defend` and `Defend Against` fields.
+- Remove conditions from the list of available traits displayed in an Item's Traits field.
 
 ### Translations
 
@@ -61,7 +63,6 @@
 - Allow an Active Effect to be active only when the owning Actor has a particular condition or (equipped) item with a particular trait.
   - For equippable Items (armor, weapons), the item must be equipped for its traits to be considered.
   - e.g. "Cloak of Darkness" effects can have "Active Only with these traits" set to "Dim" (or "Dark" or "Pitch Black") to set the appropriate bonus to toughness.
-- Allow any skill to be specified in an Active Effect's `Apply If Attack/Defend` and `Defend Against` fields.
 
 ### Bug Fixes